Logging
- Adds a handy logging macro
- Can log to a file
- Optimizations across the board:
-- Removed constantly getting a pointer to a sprite, and instead opted
for a global pointer to a sprite
-- Error sprite that loads if given sprite is invalid
- Advanced FPS tracking/capping added.

+++ source/stringify.cpp
@@ -50,7 +50,7 @@ float Stringify::ToFloat(const std::string& s)
 {
     if(s == "") return 0;
 
-    int x;
+    float x;
     std::stringstream strStream(s);
 
     strStream >> x;
@@ -62,7 +62,7 @@ double Stringify::ToDouble(const std::string& s)
 {
     if(s == "") return 0;
 
-    int x;
+    double x;
     std::stringstream strStream(s);
 
     strStream >> x;
